refactor : Move 프로젝트 구조 단순화
- 유저에 관련된 것들끼리, 로그인에 관련된 것 끼리 구조화
