+++ src/main/java/com/talearnt/login/kakao/KakaoLoginReqDTO.java
@@ -1,7 +1,5 @@
-package com.talearnt.login;
+package com.talearnt.login.kakao;
 
-import com.talearnt.enums.Gender;
-import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.*;
 
 @Builder
