+++ src/main/java/com/talearnt/user/repository/UserRepository.java
@@ -1,9 +1,8 @@
-package com.talearnt.join;
+package com.talearnt.user.repository;
 
+import com.talearnt.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
 
 
 @Repository
